requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
googlemaps==4.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
prettyprint==0.1.5
python-dotenv==0.12.0
requests==2.23.0
selenium==3.141.0
soupsieve==2.0
urllib3==1.25.8
Werkzeug==1.0.0
